#444: Moved the PassiveDebuffsResistanceParser in the model folder
 -
